Fix eating proportionally
r5245
2019-12-19

## FIX
1. Correct getting stuck and not being able to eat any
   preserved in some cases.
   - Increase the eating speed of a person.

# CHANGE
1. Rework numerous methods in EatDrink.
2. Correct the display of home town in a person's
   unit window.
3. Reword a log statment in RepairEVAMalfunction and
   RepairMalfunction.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/Inventory.java
@@ -2593,7 +2593,7 @@ public class Inventory implements Serializable {
 		double containedCapacity = 0D;
 		// Check for null
 		if (containersCapacityCacheDirty.containsKey(resource)) {
-			if (containersCapacityCacheDirty.get(resource)) {
+			if (containersCapacityCacheDirty.get((Integer)resource)) {
 				if (containedUnitIDs != null) {
 					for (Unit unit : getContainedUnits()) {
 						if (unit instanceof Container) {
